Fix warnings/lints
