+++ tests/util/mod.rs
@@ -1,0 +1,5 @@
+// Not every test module -- which are compiled at seperate binaries, use every function, causing dead_code to be emitted.
+#![allow(dead_code)]
+
+pub mod dura;
+pub mod git_repo;
